+++ Firestore/core/src/core/key_field_not_in_filter.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Google
+ * Copyright 2020 Google
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-#ifndef FIRESTORE_CORE_SRC_CORE_KEY_FIELD_IN_FILTER_H_
-#define FIRESTORE_CORE_SRC_CORE_KEY_FIELD_IN_FILTER_H_
+#ifndef FIRESTORE_CORE_SRC_CORE_KEY_FIELD_NOT_IN_FILTER_H_
+#define FIRESTORE_CORE_SRC_CORE_KEY_FIELD_NOT_IN_FILTER_H_
 
-#include <Firestore/core/src/model/document.h>
-#include <Firestore/core/src/model/field_value.h>
 #include <string>
 
 #include "Firestore/core/src/core/field_filter.h"
@@ -28,16 +26,11 @@ namespace firestore {
 namespace core {
 
 /**
- * A Filter that matches on an array of key fields.
+ * A Filter that matches on key fields not present within an array..
  */
-class KeyFieldInFilter : public FieldFilter {
+class KeyFieldNotInFilter : public FieldFilter {
  public:
-  KeyFieldInFilter(model::FieldPath field, model::FieldValue value);
-
-  static bool Contains(const model::FieldValue::Array& array_value,
-                       const model::Document& doc);
-
-  static void ValidateArrayValue(const model::FieldValue& value);
+  KeyFieldNotInFilter(model::FieldPath field, model::FieldValue value);
 
  private:
   class Rep;
@@ -47,4 +40,4 @@ class KeyFieldInFilter : public FieldFilter {
 }  // namespace firestore
 }  // namespace firebase
 
-#endif  // FIRESTORE_CORE_SRC_CORE_KEY_FIELD_IN_FILTER_H_
+#endif  // FIRESTORE_CORE_SRC_CORE_KEY_FIELD_NOT_IN_FILTER_H_