+++ Firestore/core/test/unit/remote/serializer_test.cc
@@ -369,9 +369,10 @@ class SerializerTest : public ::testing::Test {
     create_time_proto->set_nanos(4321);
   }
 
-  void ExpectUnaryOperator(const FieldValue& value,
+  void ExpectUnaryOperator(std::string opStr,
+                           const FieldValue& value,
                            v1::StructuredQuery::UnaryFilter::Operator op) {
-    core::Query q = Query("docs").AddingFilter(Filter("prop", "==", value));
+    core::Query q = Query("docs").AddingFilter(Filter("prop", opStr, value));
     TargetData model = CreateTargetData(std::move(q));
 
     v1::Target proto;
@@ -383,6 +384,15 @@ class SerializerTest : public ::testing::Test {
     *proto.mutable_query()->mutable_structured_query()->add_from() =
         std::move(from);
 
+    // Add extra ORDER_BY field for '!=' since it is an inequality.
+    if (opStr == "!=") {
+      v1::StructuredQuery::Order order1;
+      order1.mutable_field()->set_field_path("prop");
+      order1.set_direction(v1::StructuredQuery::ASCENDING);
+      *proto.mutable_query()->mutable_structured_query()->add_order_by() =
+          std::move(order1);
+    }
+
     v1::StructuredQuery::Order order;
     order.mutable_field()->set_field_path(FieldPath::kDocumentKeyPath);
     order.set_direction(v1::StructuredQuery::ASCENDING);
@@ -393,6 +403,7 @@ class SerializerTest : public ::testing::Test {
                                                     ->mutable_structured_query()
                                                     ->mutable_where()
                                                     ->mutable_unary_filter();
+
     filter.mutable_field()->set_field_path("prop");
     filter.set_op(op);
 
@@ -1335,14 +1346,26 @@ TEST_F(SerializerTest, EncodesMultipleFiltersOnDeeperCollections) {
 
 TEST_F(SerializerTest, EncodesNullFilter) {
   SCOPED_TRACE("EncodesNullFilter");
-  ExpectUnaryOperator(Value(nullptr),
+  ExpectUnaryOperator("==", Value(nullptr),
                       v1::StructuredQuery_UnaryFilter_Operator_IS_NULL);
 }
 
 TEST_F(SerializerTest, EncodesNanFilter) {
   SCOPED_TRACE("EncodesNanFilter");
-  ExpectUnaryOperator(Value(NAN),
+  ExpectUnaryOperator("==", Value(NAN),
                       v1::StructuredQuery_UnaryFilter_Operator_IS_NAN);
+}
+
+TEST_F(SerializerTest, EncodesNotNullFilter) {
+  SCOPED_TRACE("EncodesNotNullFilter");
+  ExpectUnaryOperator("!=", Value(nullptr),
+                      v1::StructuredQuery_UnaryFilter_Operator_IS_NOT_NULL);
+}
+
+TEST_F(SerializerTest, EncodesNotNanFilter) {
+  SCOPED_TRACE("EncodesNotNanFilter");
+  ExpectUnaryOperator("!=", Value(NAN),
+                      v1::StructuredQuery_UnaryFilter_Operator_IS_NOT_NAN);
 }
 
 TEST_F(SerializerTest, EncodesSortOrders) {
@@ -1890,6 +1913,18 @@ TEST_F(SerializerTest, EncodesFieldFilter) {
   ExpectRoundTrip(model, proto);
 }
 
+TEST_F(SerializerTest, EncodesNotEqualFilter) {
+  auto model = testutil::Filter("item.tags", "!=", "food");
+
+  v1::StructuredQuery::Filter proto;
+  v1::StructuredQuery::FieldFilter& field = *proto.mutable_field_filter();
+  field.mutable_field()->set_field_path("item.tags");
+  field.set_op(v1::StructuredQuery::FieldFilter::NOT_EQUAL);
+  *field.mutable_value() = ValueProto("food");
+
+  ExpectRoundTrip(model, proto);
+}
+
 TEST_F(SerializerTest, EncodesArrayContainsFilter) {
   auto model = testutil::Filter("item.tags", "array_contains", "food");
 
@@ -1923,6 +1958,32 @@ TEST_F(SerializerTest, EncodesInFilter) {
   field.mutable_field()->set_field_path("item.tags");
   field.set_op(v1::StructuredQuery::FieldFilter::IN_);
   *field.mutable_value() = ValueProto(std::vector<FieldValue>{Value("food")});
+
+  ExpectRoundTrip(model, proto);
+}
+
+TEST_F(SerializerTest, EncodesNotInFilter) {
+  auto model = testutil::Filter("item.tags", "not-in", Array("food"));
+
+  v1::StructuredQuery::Filter proto;
+  v1::StructuredQuery::FieldFilter& field = *proto.mutable_field_filter();
+  field.mutable_field()->set_field_path("item.tags");
+  field.set_op(v1::StructuredQuery::FieldFilter::NOT_IN);
+  *field.mutable_value() = ValueProto(std::vector<FieldValue>{Value("food")});
+
+  ExpectRoundTrip(model, proto);
+}
+
+TEST_F(SerializerTest, EncodesNotInFilterWithNull) {
+  auto model =
+      testutil::Filter("item.tags", "not-in", Array(FieldValue::Null()));
+
+  v1::StructuredQuery::Filter proto;
+  v1::StructuredQuery::FieldFilter& field = *proto.mutable_field_filter();
+  field.mutable_field()->set_field_path("item.tags");
+  field.set_op(v1::StructuredQuery::FieldFilter::NOT_IN);
+  *field.mutable_value() =
+      ValueProto(std::vector<FieldValue>{FieldValue::Null()});
 
   ExpectRoundTrip(model, proto);
 }