+++ Firestore/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Unreleased
+- [feature] Added `whereField(_:notIn:)` and `whereField(_:isNotEqualTo:)` query
+operators. `whereField(_:notIn:)` finds documents where a specified field’s
+value is not in a specified array. `whereField(_:isNotEqualTo:)` finds documents
+where a specified field's value does not equal the specified value.
 
 # v1.17.1
 - [fixed] Fix gRPC documentation warning surfaced in Xcode (#6340).